requirements.txt
fastapi==0.115.2
uvicorn[standard]
python-multipart
python-jose[cryptography]
passlib[bcrypt]
google-generativeai==0.8.3
markdown
jwt
sqlmodel
httpx
PyJWT